Use pushInsteadOf to always push with ssh
- This is preferred over always using ssh because this allows an
engineer to pull using https (which they often do with open source
repositories), while always using their ssh key to push (in the case of
both open source and private repositories). This also maintains the
functionality of pulling with ssh if the repository was cloned with the
ssh url in the case of private repositories.