Fix ssh agent forwarding
- ssh agent forwarding allows an engineer to forward their local agent
to the server so they can use a locally loaded key e.g ssh -A
- this is discovered by checking if the SSH_AUTH_SOCK environment
variable is set. If it is set when the .bash_profile is loaded, then
that probably means the user ssh'd onto the machine with a ssh agent
forwarded and we shouldn't override that agent with our own. If it is
not set, then it is probably safe to set it to the single SSH_AUTH_SOCK.
Best of both worlds.